+++ unit/test/Invoke-VSTeamRequest.Tests.ps1
@@ -11,6 +11,7 @@ $sut = (Split-Path -Leaf $MyInvocation.MyCommand.Path).Replace(".Tests.", ".")
 . "$here/../../Source/Classes/VSTeamLeaf.ps1"
 . "$here/../../Source/Classes/VSTeamDirectory.ps1"
 . "$here/../../Source/Classes/VSTeamProjectCache.ps1"
+. "$here/../../Source/Classes/InvokeCompleter.ps1"
 . "$here/../../Source/Classes/ProjectCompleter.ps1"
 . "$here/../../Source/Classes/ProjectValidateAttribute.ps1"
 . "$here/../../Source/Classes/UncachedProjectCompleter.ps1"